+++ postgres/lib/neon_postgres.rb
@@ -3,6 +3,7 @@ require "sequel"
 require "neon_postgres/version"
 require "neon_postgres/database"
 require "neon_postgres/document_sanitizer"
+require "neon_postgres/sequel_models"
 Dir["#{__dir__}/neon_postgres/inter_database_copy/*"].sort.each { |file|
   next if /copy.rb$/.match?(file)
   require file
