+++ postgres/lib/neon_postgres/version.rb
@@ -1,3 +1,3 @@
 module NeonPostgres
-  VERSION = "0.0.1"
+  VERSION = "0.0.2"
 end
