+++ postgres/spec/fixtures/documents.yml.erb
@@ -1,10 +1,10 @@
 filed_summons:
   filename: Summons
-  document_template: filed_document (DocumentTemplate)
+  document_template: filed_document (NeonPostgres::Models::DocumentTemplate)
   documentable_table_name: matter_document
   gcp_url: <%= "https://neon-law-staging-private-assets.storage.googleapis.com/matters/#{Faker::Internet.uuid}" %>
 filed_complaint:
   filename: Complaint
-  document_template: filed_document (DocumentTemplate)
+  document_template: filed_document (NeonPostgres::Models::DocumentTemplate)
   documentable_table_name: matter_document
   gcp_url: <%= "https://neon-law-staging-private-assets.storage.googleapis.com/matters/#{Faker::Internet.uuid}" %>
