+++ postgres/spec/spec_helper.rb
@@ -6,8 +6,11 @@ require "fixture_dependencies/rspec/sequel"
 
 Sequel::Model.db = NeonPostgres::Database.connection
 FixtureDependencies.fixture_path = "#{__dir__}/fixtures"
-
-require_relative "sequel_models"
+FixtureDependencies.class_map[:person] = NeonPostgres::Models::Person
+FixtureDependencies.class_map[:document] = NeonPostgres::Models::Document
+FixtureDependencies.class_map[:document_template] = NeonPostgres::Models::DocumentTemplate
+FixtureDependencies.class_map[:matter_template] = NeonPostgres::Models::MatterTemplate
+FixtureDependencies.class_map[:matter] = NeonPostgres::Models::Matter
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
